+++ tests/codegen/map_increment_decrement.cpp
@@ -33,14 +33,14 @@ entry:
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %2)
   store i64 10, i64* %"@x_val", align 8
   %pseudo = tail call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %update_elem = call i64 inttoptr (i64 2 to i64 (i8*, i8*, i8*, i64)*)(i64 %pseudo, i64* nonnull %"@x_key", i64* nonnull %"@x_val", i64 0)
+  %update_elem = call i64 inttoptr (i64 2 to i64 (i8*, i64*, i64*, i64)*)(i64 %pseudo, i64* nonnull %"@x_key", i64* nonnull %"@x_val", i64 0)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %1)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %2)
   %3 = bitcast i64* %"@x_key1" to i8*
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %3)
   store i64 0, i64* %"@x_key1", align 8
   %pseudo2 = call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %lookup_elem = call i8* inttoptr (i64 1 to i8* (i8*, i8*)*)(i64 %pseudo2, i64* nonnull %"@x_key1")
+  %lookup_elem = call i8* inttoptr (i64 1 to i8* (i8*, i64*)*)(i64 %pseudo2, i64* nonnull %"@x_key1")
   %map_lookup_cond = icmp eq i8* %lookup_elem, null
   br i1 %map_lookup_cond, label %lookup_merge, label %lookup_success
 
@@ -55,14 +55,14 @@ lookup_merge:                                     ; preds = %entry, %lookup_succ
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %5)
   store i64 %lookup_elem_val.0, i64* %"@x_newval", align 8
   %pseudo3 = call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %update_elem4 = call i64 inttoptr (i64 2 to i64 (i8*, i8*, i8*, i64)*)(i64 %pseudo3, i64* nonnull %"@x_key1", i64* nonnull %"@x_newval", i64 0)
+  %update_elem4 = call i64 inttoptr (i64 2 to i64 (i8*, i64*, i64*, i64)*)(i64 %pseudo3, i64* nonnull %"@x_key1", i64* nonnull %"@x_newval", i64 0)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %3)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %5)
   %6 = bitcast i64* %"@x_key5" to i8*
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %6)
   store i64 0, i64* %"@x_key5", align 8
   %pseudo6 = call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %lookup_elem7 = call i8* inttoptr (i64 1 to i8* (i8*, i8*)*)(i64 %pseudo6, i64* nonnull %"@x_key5")
+  %lookup_elem7 = call i8* inttoptr (i64 1 to i8* (i8*, i64*)*)(i64 %pseudo6, i64* nonnull %"@x_key5")
   %map_lookup_cond12 = icmp eq i8* %lookup_elem7, null
   br i1 %map_lookup_cond12, label %lookup_merge10, label %lookup_success8
 
@@ -77,14 +77,14 @@ lookup_merge10:                                   ; preds = %lookup_merge, %look
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %8)
   store i64 %lookup_elem_val11.0, i64* %"@x_newval13", align 8
   %pseudo14 = call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %update_elem15 = call i64 inttoptr (i64 2 to i64 (i8*, i8*, i8*, i64)*)(i64 %pseudo14, i64* nonnull %"@x_key5", i64* nonnull %"@x_newval13", i64 0)
+  %update_elem15 = call i64 inttoptr (i64 2 to i64 (i8*, i64*, i64*, i64)*)(i64 %pseudo14, i64* nonnull %"@x_key5", i64* nonnull %"@x_newval13", i64 0)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %6)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %8)
   %9 = bitcast i64* %"@x_key16" to i8*
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %9)
   store i64 0, i64* %"@x_key16", align 8
   %pseudo17 = call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %lookup_elem18 = call i8* inttoptr (i64 1 to i8* (i8*, i8*)*)(i64 %pseudo17, i64* nonnull %"@x_key16")
+  %lookup_elem18 = call i8* inttoptr (i64 1 to i8* (i8*, i64*)*)(i64 %pseudo17, i64* nonnull %"@x_key16")
   %map_lookup_cond23 = icmp eq i8* %lookup_elem18, null
   br i1 %map_lookup_cond23, label %lookup_merge21, label %lookup_success19
 
@@ -99,14 +99,14 @@ lookup_merge21:                                   ; preds = %lookup_merge10, %lo
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %11)
   store i64 %lookup_elem_val22.0, i64* %"@x_newval24", align 8
   %pseudo25 = call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %update_elem26 = call i64 inttoptr (i64 2 to i64 (i8*, i8*, i8*, i64)*)(i64 %pseudo25, i64* nonnull %"@x_key16", i64* nonnull %"@x_newval24", i64 0)
+  %update_elem26 = call i64 inttoptr (i64 2 to i64 (i8*, i64*, i64*, i64)*)(i64 %pseudo25, i64* nonnull %"@x_key16", i64* nonnull %"@x_newval24", i64 0)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %9)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %11)
   %12 = bitcast i64* %"@x_key27" to i8*
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %12)
   store i64 0, i64* %"@x_key27", align 8
   %pseudo28 = call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %lookup_elem29 = call i8* inttoptr (i64 1 to i8* (i8*, i8*)*)(i64 %pseudo28, i64* nonnull %"@x_key27")
+  %lookup_elem29 = call i8* inttoptr (i64 1 to i8* (i8*, i64*)*)(i64 %pseudo28, i64* nonnull %"@x_key27")
   %map_lookup_cond34 = icmp eq i8* %lookup_elem29, null
   br i1 %map_lookup_cond34, label %lookup_merge32, label %lookup_success30
 
@@ -121,7 +121,7 @@ lookup_merge32:                                   ; preds = %lookup_merge21, %lo
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %14)
   store i64 %lookup_elem_val33.0, i64* %"@x_newval35", align 8
   %pseudo36 = call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %update_elem37 = call i64 inttoptr (i64 2 to i64 (i8*, i8*, i8*, i64)*)(i64 %pseudo36, i64* nonnull %"@x_key27", i64* nonnull %"@x_newval35", i64 0)
+  %update_elem37 = call i64 inttoptr (i64 2 to i64 (i8*, i64*, i64*, i64)*)(i64 %pseudo36, i64* nonnull %"@x_key27", i64* nonnull %"@x_newval35", i64 0)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %12)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %14)
   ret i64 0
